LayoutTests/imported/w3c/web-platform-tests/dom/events/scrolling/wheel-event-transactions-basic_include=target-basic-expected.txt


FAIL Wheel events should be captured to one target #1 assert_equals: Wheel event at index `0` does not have the expected target expected Element node <div id="firstRootSpacer" class="spacer" style="backgroun... but got Element node <div id="secondRootSpacer" class="spacer" style="backgrou...
